pyproject.toml
[project]
name = "gridbot"
version = "0.1.0"
description = "A minimal Python Spot grid bot MVP running locally"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
